The Resume functionality was enabled in the following ways: 1. CountDownTimerViewModel.kt A new member attribute `remainingDurationInMilliSeconds` was created to keep track of the milliseconds by updating its value in `updateDuration` method.
The new attribute was then used to create a new TicToc timer in `resumeCountDownTimer()`

2. TicTocTimer.kt
It was noted that after resuming, the tic had a consistent lag of 1000 milliseconds. This was because the 'remainingDurationInMilliSeconds' attribute was being updated only after every 1000 milliseconds.
The issue was rectified by updating the onTick interval to 100 milliseconds. The result was that 'remainingDurationInMilliSeconds' was updated every 100 milliseconds thus reducing the margin of error.
No side effects occurred in 'updateDuration' because the format utility still did its job correctly.